requirements.txt
pandas
pillow 
opencv-python
numpy
matplotlib
bs4
requests
tifffile
torch